Isolate `inflictorWithValue`
