+++ src/components/Visualizations/index.js
@@ -3,3 +3,4 @@ export { default as KDA } from './Table/KDA';
 export { default as TableHeroImage } from './Table/HeroImage';
 export { default as FromNowTooltip } from './FromNowTooltip';
 export { default as Graph, TrendGraph, HistogramGraph, MMRGraph } from './Graph';
+export { default as inflictorWithValue } from './inflictorWithValue';
