+++ src/components/Visualizations/inflictorWithValue.css
@@ -1,0 +1,33 @@
+@import "../palette.css";
+
+.inflictorWithValue {
+  position: relative;
+  float: left;
+  margin: 2px;
+  height: 29px;
+
+  & .overlay {
+    background-color: var(--darkPrimaryColor);
+    font-size: 10px;
+    position: absolute;
+    left: 0;
+    bottom: 0;
+    width: 100%;
+    text-align: center;
+  }
+  & img {
+    height: 29px;
+  }
+
+  & .tooltip {
+    & > div:matches(::after, ::before) {
+      content: none !important;
+    }
+    & > div {
+      background-color: var(--darkPrimaryColor);
+      color: var(--textColorPrimary);
+      max-width: 300px;
+      white-space: pre-wrap;
+    }
+  }
+}
